Backend/requirements.txt
fastapi
uvicorn
firebase-admin
pyjwt
bcrypt
python-multipart
